Adjustments and bug fixes to BatchedUpdateAggregator
